% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hdr.R
\name{instance_hdr}
\alias{instance_hdr}
\title{Simulation for high-dimensional linear regression model variable selection}
\usage{
instance_hdr(
  n,
  p,
  s0,
  xtype = c("toeplitz", "exp.decay", "equi.corr"),
  btype = "U[-2,2]",
  permuted = TRUE,
  x.par = switch(xtype, toeplitz = 1/3, equi.corr = 1/20, exp.decay = c(0.4, 5)),
  yfun,
  yargs = NULL,
  fitfun,
  fitargs = NULL,
  postfun,
  postargs = NULL
)
}
\arguments{
\item{n}{Sample size, number of rows in design matrix.}

\item{p}{Number of random predictor variables or columns in design matrix.}

\item{s0}{Sparsity or number of nonzero coefficients in the true linear model.}

\item{xtype, btype, permuted, x.par}{Linear model data generation parameters, see \link[hdi]{rXb} for details.}

\item{yfun, yargs}{Function and (optional) arguments for generating outcome variable.}

\item{fitfun, fitargs}{Function and (optional) arguments for fitting models to simulated data.}

\item{postfun, postargs}{Function and (optional) arguments for post-processing fitted models.}
}
\value{
Outputs from \code{postfun} after applying it to the model fitted by \code{fitfun} on one instance of simulated data.
}
\description{
\code{this} repeats \code{that} multiple times.
}
\details{
After generating simulated data this calls \code{something} for inference.
}
\examples{
n <- 100
p <- 200
s0 <- 5
}
