% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hdr.R
\name{instance_hdr}
\alias{instance_hdr}
\title{instance_hdr
Simulation for high-dimensional linear regression models.}
\usage{
instance_hdr(
  n,
  p,
  s0,
  xtype = c("toeplitz", "exp.decay", "equi.corr"),
  btype = "U[-2,2]",
  permuted = TRUE,
  x.par = switch(xtype, toeplitz = 1/3, equi.corr = 1/20, exp.decay = c(0.4, 5)),
  yfun = y_standard_linear,
  yargs = NULL,
  fitfun = fit_glmnet_cvmin,
  fitargs = NULL,
  postfun = post_glmnet_cvmin,
  postargs = NULL
)
}
\arguments{
\item{n}{Sample size, number of rows in design matrix.}

\item{p}{Number of random predictor variables or columns in design matrix.}

\item{s0}{Sparsity or number of nonzero coefficients in the true linear model.}

\item{xtype, btype, permuted, x.par}{Linear model data generation parameters, see \link[hdi]{rXb} for details.}

\item{yfun, yargs}{Function and (optional) arguments for generating outcome variable.}

\item{fitfun, fitargs}{Function and (optional) arguments for fitting models to simulated data.}

\item{postfun, postargs}{Function and (optional) arguments for post-processing fitted models.}
}
\value{
Outputs from \code{postfun} after applying it to the model fitted by \code{fitfun} on one instance of simulated data.
}
\description{
Simulates data from a high-dimensional linear model and then applies user-specified functions to fit a model and process the results.
}
\details{
\enumerate{
\item Generates a random design matrix and coefficient vector using \code{\link[hdi:rXb]{hdi::rXb()}}.
\item Generates an outcome variable using \code{yfun}.
\item Fits a model to the data using the method specified with \code{fitfun}.
\item Processes the resulting model with \code{postfun}.
}
}
\examples{
\dontrun{
instance_hdr(n = 100, p = 200, s0 = 5)
}
}
