% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hdr.R
\name{simmary_coefs}
\alias{simmary_coefs}
\title{simmary_coefs
Compute simulation summary using true and estimated linear model coefficients.}
\usage{
simmary_coefs(list_of_instances)
}
\arguments{
\item{list_of_instances}{The output of a simulation function.}
}
\value{
A \code{data.frame} with one row for each simulation instance and several performance metrics:
\itemize{
\item \code{screened} Boolean indicating whether the selected model's support is a superset of the true support, i.e. if the selected variables include all the ones with nonzero coefficients in the true simulated model.
\item \code{mse} The mean-square error of estimating the true coefficients.
\item \code{beta_min} The smallest nonzero coefficient in the true model (in absolute value).
}
}
\description{
Compute simulation summary using true and estimated linear model coefficients.
}
\examples{
\dontrun{
instances <- simulate_hdr(niters = 500, n = 100, p = 200, s0 = 5)
simmary_coefs(instances)
}
}
