% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hdr.R
\name{simulate_hdr}
\alias{simulate_hdr}
\title{Simulation for high-dimensional linear regression model variable selection}
\usage{
simulate_hdr(
  niters,
  n,
  p,
  s0,
  xtype = c("toeplitz", "exp.decay", "equi.corr"),
  btype = "U[-2,2]",
  permuted = TRUE,
  x.par = switch(xtype, toeplitz = 1/3, equi.corr = 1/20, exp.decay = c(0.4, 5)),
  yfun = y_standard_linear,
  yargs = NULL,
  fitfun = fit_glmnet_cvmin,
  fitargs = NULL,
  postfun = post_glmnet_cvmin,
  postargs = NULL,
  cores = NULL
)
}
\arguments{
\item{niters}{Number of simulation iterations.}

\item{n}{Sample size, number of rows in design matrix.}

\item{p}{Number of random predictor variables or columns in design matrix.}

\item{s0}{Sparsity or number of nonzero coefficients in the true linear model.}

\item{xtype, btype, permuted, x.par}{Linear model data generation parameters, see \link[hdi]{rXb} for details.}

\item{yfun, yargs}{Function and (optional) arguments for generating outcome variable.}

\item{fitfun, fitargs}{Function and (optional) arguments for fitting models to simulated data.}

\item{postfun, postargs}{Function and (optional) arguments for post-processing fitted models.}

\item{cores}{Number of cores for parallel computation, passed to \link[parallel]{makeCluster}. Defaults to half of \link[parallel]{detectCores} when not specified.}
}
\value{
List of outputs from \code{postfun} for each of \code{niters} instances.
}
\description{
\code{this} repeats \code{that} multiple times.
}
\details{
After generating simulated data this calls \code{something} for inference.
}
\examples{
n <- 100
p <- 200
s0 <- 5
}
